Only the creator of the book can edit and delete it

+++ src/main/java/hu/unideb/inf/BookShop/controller/BookController.java
@@ -1,25 +1,31 @@
 package hu.unideb.inf.BookShop.controller;
 
+import hu.unideb.inf.BookShop.service.AuthenticationService;
 import hu.unideb.inf.BookShop.service.BookService;
 import hu.unideb.inf.BookShop.service.dto.BookDto;
+import hu.unideb.inf.BookShop.service.impl.BookServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
+import org.springframework.security.core.Authentication;
 import java.util.List;
 
 @RestController
 @RequestMapping("/api/books")
 public class BookController {
 
-    private final BookService bookService;
+    @Autowired
+    BookService bookService;
 
     @Autowired
-    public BookController(BookService bookService) {
-        this.bookService = bookService;
-    }
+    AuthenticationService authenticationService;
+
 
     @PostMapping
-    public BookDto createBook(@RequestBody BookDto bookDto) {
+    public BookDto createBook(@RequestBody BookDto bookDto, Authentication authentication) {
+        String email = authentication.getName();
+        bookDto.setCreatorEmail(email);
         return bookService.save(bookDto);
     }
 
@@ -34,12 +40,29 @@ public class BookController {
     }
 
     @PutMapping("/{id}")
-    public BookDto updateBook(@PathVariable Long id, @RequestBody BookDto bookDto) {
-        return bookService.update(id, bookDto);
+    public ResponseEntity<BookDto> updateBook(@PathVariable Long id, @RequestBody BookDto bookDto, Authentication authentication) {
+        BookDto existingBook = bookService.findById(id);
+
+        if (!existingBook.getCreatorEmail().equals(authentication.getName())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+        bookDto.setId(id);
+        return ResponseEntity.ok(bookService.update(id, bookDto));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBook(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBook(@PathVariable Long id, Authentication authentication) {
+        BookDto existingBook = bookService.findById(id);
+        String creatorEmail = existingBook.getCreatorEmail();
+        String currentUserEmail = authentication.getName();
+
+        if (!creatorEmail.equals(currentUserEmail)) {
+            System.out.println("Forbidden: " + currentUserEmail + " is not the creator of this book.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         bookService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
